Time Entity 이름 수정 및 상속관계 생성 & Family 엔티티 구현
