--- src/main/kotlin/com/record/zooc/domain/entity/BaseTimeEntityCreated.kt
@@ -1,18 +1,0 @@
-package com.record.zooc.domain.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.MappedSuperclass
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-abstract class BaseTimeEntityCreated {
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    var createdAt: LocalDateTime = LocalDateTime.now()
-        protected set
-    // var로 해야지 시간 저장될 때 오류 안남
-}
