+++ src/main/kotlin/com/record/zooc/domain/entity/Family.kt
@@ -1,10 +1,19 @@
-package com.record.zooc.domain.entity
+import com.record.zooc.domain.entity.time.ModifiedTimeEntity
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
 
 @Entity
-class Family {
-
+class Family(
+    code: String,
+) : ModifiedTimeEntity() {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = 0
         protected set
+
+    @Column
+    val code: String = code
 }
