+++ src/main/kotlin/com/record/zooc/domain/entity/Pet.kt
@@ -14,7 +14,6 @@ class Pet(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null
-        protected set
 
     @Column
     var name: String = name
