+++ src/main/kotlin/com/record/zooc/domain/entity/time/CreatedTimeEntity.kt
@@ -1,26 +1,20 @@
-package com.record.zooc.domain.entity
+package com.record.zooc.domain.entity.time
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import jakarta.persistence.Column
 import jakarta.persistence.EntityListeners
 import jakarta.persistence.MappedSuperclass
 import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import java.time.LocalDateTime
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener::class)
-abstract class BaseTimeEntityModified {
+abstract class CreatedTimeEntity {
     @CreatedDate
     @Column(nullable = false, updatable = false)
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     var createdAt: LocalDateTime = LocalDateTime.now()
         protected set
-
-    @LastModifiedDate
-    @Column(nullable = false)
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    var modifiedAt: LocalDateTime = LocalDateTime.now()
-        protected set
+    // var로 해야지 시간 저장될 때 오류 안남
 }
