+++ src/main/kotlin/com/record/zooc/domain/entity/time/ModifiedTimeEntity.kt
@@ -1,0 +1,21 @@
+package com.record.zooc.domain.entity.time
+
+import com.fasterxml.jackson.annotation.JsonFormat
+import com.record.zooc.domain.entity.time.CreatedTimeEntity
+import jakarta.persistence.Column
+import jakarta.persistence.EntityListeners
+import jakarta.persistence.MappedSuperclass
+import org.springframework.data.annotation.LastModifiedDate
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener::class)
+abstract class ModifiedTimeEntity : CreatedTimeEntity() {
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    var modifiedAt: LocalDateTime = LocalDateTime.now()
+        protected set
+}
